Damn I really got to start doing more commits.
Added HTML to it.
Added a progress.md so I can have something to more closely monitor my progress but also to take more notes and start reveiwing my work.
Now there's date filtering; Three different pages, one for past and present, another for present and future, and another for all assignments.
There's also a function that organizes everything by due date from the earliest date to the most in the future I guess that's how you say that.....
Added button for the links, and a basic layout that is representative of the content that I am using in order to add assignments to Google Calendar.
Honestly don't even know the hell I am doing next.