Added a few things to progress.md
Basically reviewing my work and making sure I understand the things I am working with. Reading over documentation for Google API use. So far it's been challenging but happy to say that I've use ChatGPT like once so far and it was only so it could pull up resources so I could speed up the process of finding actually useful information. That is something that I need to work on myself and just get better at finding information and applying the concepts that I am reading. So far my main challenge is that I am horrible at just making quick examples so I can practice a bit but it's whatever I guess.